Implementing persistence_matrix instead of Ripser
The code is set up to calculate the persistence set D_{n,k} with n=2k+2, so we use our custom algorithm to find the VR persistence. The advantage over using Ripser is that we don't have to write a distance matrix to disk in every iteration. This might be fixed in a future commit.